chore: move Initialization screen to component
